srcs/user_management_api/requirements.txt
asgiref==3.8.1
Django==5.0.6
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
logstash-formatter==0.5.17
logstash-logger==0.1.0
Markdown==3.6
packaging==24.1
psycopg==3.1.19
PyJWT==2.8.0
python-logstash==0.4.8
sqlparse==0.5.0
typing_extensions==4.12.2
